+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+* Handle case where HTTP response is not valid JSON.
+
+
 ## [0.2.0] - 2022-09-26
 ### Added
 * Add new `Formality` options: `PreferLess` and `PreferMore`.
@@ -40,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-java/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/DeepLcom/deepl-java/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/DeepLcom/deepl-java/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/DeepLcom/deepl-java/compare/v0.1.1...v0.1.2